+++ src/egeo-demo/st-table-demo/st-table-demo.component.ts
@@ -72,4 +72,10 @@ export class StTableDemoComponent {
          completedProfile: '80%'
       }
    ];
+
+   public selected: boolean[] = [];
+
+   public onSelectRow(event: any, rowIndex: number) {
+      this.selected[rowIndex] = event.checked;
+   }
 }