+++ src/egeo-demo/st-table-demo/st-table-demo.module.ts
@@ -10,7 +10,7 @@
  */
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { StDemoGeneratorModule, StTableModule, StButtonModule } from '@stratio/egeo';
+import { StDemoGeneratorModule, StTableModule, StButtonModule, StCheckboxModule } from '@stratio/egeo';
 
 import { StTableDemoComponent } from './st-table-demo.component';
 
@@ -19,6 +19,7 @@ import { StTableDemoComponent } from './st-table-demo.component';
       CommonModule,
       StTableModule,
       StButtonModule,
+      StCheckboxModule,
       StDemoGeneratorModule.withComponents({ components: [StTableDemoComponent] })
    ],
    declarations: [StTableDemoComponent]